feat: v0.1.6 — çoklu barkod sistemi, fiyat teklifi, güncelleme sistemi, bug fix'ler
Yeni özellikler:
- Çoklu barkod (alias) sistemi: bir ürüne ana barkod dışında ek barkodlar
  tanımlanabilir (product_barcodes tablosu, 5 yeni endpoint, ürün formunda
  barkod havuzu yönetimi, POS/Stoklar tarama-arama entegrasyonu, Excel
  içe/dışa aktarımda Alternatif Barkodlar sütunu, barkod çakışması çözümü)
- Fiyat teklifi PDF'i: satış yapmadan sepetten bilgi amaçlı PDF oluşturma
- Sunucu bazlı güncelleme sistemi: admin min_app_version belirler, yetersiz
  sürümdeki istemciler engellenir ve GitHub releases'e yönlendirilir
  (yarım kalmış önceki implementasyon tamamlandı)

Bug fix'ler:
- Sepette/stokta artık her ürün için ondalık miktar girilebiliyor (0.5 kg vb.)
- Barkod tarama: çerçeve dışı artık algılanmıyor, 3 saniyelik "Taranıyor"
  onay bekleme süresi eklendi (yanlış okumaları önler)
- Sepet miktar/indirim dialoglarında klavye açılınca oluşan taşma giderildi

+++ inventra_app/lib/features/product/providers/product_provider.dart
@@ -133,7 +133,7 @@ class ProductNotifier extends StateNotifier<AsyncValue<List<Product>>> {
     String barcode,
     String name,
     double salePrice,
-    int stock, {
+    double stock, {
     double purchasePrice = 0,
     double vatRate = 20,
     String unit = "Adet",
@@ -261,7 +261,7 @@ class ProductNotifier extends StateNotifier<AsyncValue<List<Product>>> {
     }
   }
 
-  Future<int> bulkAddStock(Map<String, int> productStocks) async {
+  Future<int> bulkAddStock(Map<String, double> productStocks) async {
     try {
       final items = productStocks.entries.map((e) => {
         'product_id': e.key,
@@ -279,7 +279,7 @@ class ProductNotifier extends StateNotifier<AsyncValue<List<Product>>> {
     }
   }
 
-  Future<int> bulkSubtractStock(Map<String, int> productStocks) async {
+  Future<int> bulkSubtractStock(Map<String, double> productStocks) async {
     try {
       final items = productStocks.entries.map((e) => {
         'product_id': e.key,
